Cleaning code and splitting them into multiple files. Implementation of check completed
