+++ airbyte-integrations/connectors/source-snowflake/source_snowflake/source.py
@@ -4,293 +4,71 @@
 import uuid
 from abc import ABC
 from typing import Any, Iterable, List, Mapping, MutableMapping, Optional, Tuple, Union
-
+from .streams import SnowflakeStream, CheckConnectionStream
 import requests
 from airbyte_cdk.sources import AbstractSource
 from airbyte_cdk.sources.streams import Stream
-from airbyte_cdk.sources.streams.http import HttpStream
 from airbyte_cdk.sources.streams.http.requests_native_auth import TokenAuthenticator
 from airbyte_protocol.models import SyncMode
 
 from .authenticator import SnowflakeJwtAuthenticator
 
-"""
-TODO: Most comments in this class are instructive and should be deleted after the source is implemented.
-
-This file provides a stubbed example of how to use the Airbyte CDK to develop both a source connector which supports full refresh or and an
-incremental syncs from an HTTP API.
-
-The various TODOs are both implementation hints and steps - fulfilling all the TODOs should be sufficient to implement one basic and one incremental
-stream from a source. This pattern is the same one used by Airbyte internally to implement connectors.
-
-The approach here is not authoritative, and devs are free to use their own judgement.
-
-There are additional required TODOs in the files within the integration_tests folder and the spec.yaml file.
-"""
-
-
-# Basic full refresh stream
-class SnowflakeStream(HttpStream, ABC):
-    """
-    TODO remove this comment
-
-    This class represents a stream output by the connector.
-    This is an abstract base class meant to contain all the common functionality at the API level e.g: the API base URL, pagination strategy,
-    parsing responses etc..
-
-    Each stream should extend this class (or another abstract subclass of it) to specify behavior unique to that stream.
-
-    Typically for REST APIs each stream corresponds to a resource in the API. For example if the API
-    contains the endpoints
-        - GET v1/customers
-        - GET v1/employees
-
-    then you should have three classes:
-    `class SnowflakeStream(HttpStream, ABC)` which is the current class
-    `class Customers(SnowflakeStream)` contains behavior to pull data for customers using v1/customers
-    `class Employees(SnowflakeStream)` contains behavior to pull data for employees using v1/employees
-
-    If some streams implement incremental sync, it is typical to create another class
-    `class IncrementalSnowflakeStream((SnowflakeStream), ABC)` then have concrete stream implementations extend it. An example
-    is provided below.
-
-    See the reference docs for the full list of configurable options.
-    """
-    url_suffix = "api/v2/statements"
-    url_base = ""
-
-    @property
-    def statement(self):
-        return None
-
-    def next_page_token(self, response: requests.Response) -> Optional[Mapping[str, Any]]:
-        return None
-
-    def request_params(
-            self, stream_state: Mapping[str, Any], stream_slice: Mapping[str, any] = None, next_page_token: Mapping[str, Any] = None
-    ) -> MutableMapping[str, Any]:
-        """
-        TODO: Override this method to define any query parameters to be set. Remove this method if you don't need to define request params.
-        Usually contains common params e.g. pagination size etc.
-        """
-        params = {
-            "requestId": "96c46822-fec9-43ee-accf-b2dd222bd024",#uuid.uuid4(),
-            "async": "false"
-        }
-        return params
-
-    @property
-    def http_method(self) -> str:
-        return "POST"
-
-    def request_headers(
-            self,
-            stream_state: Optional[Mapping[str, Any]],
-            stream_slice: Optional[Mapping[str, Any]] = None,
-            next_page_token: Optional[Mapping[str, Any]] = None,
-    ) -> Mapping[str, Any]:
-        headers = {
-            'User-Agent': 'myApplication/1.0',
-            'X-Snowflake-Authorization-Token-Type': 'KEYPAIR_JWT',
-            'Content-Type': 'application/json',
-            'Accept': 'application/json',
-        }
-
-        return headers
-
-    def parse_response(self, response: requests.Response, **kwargs) -> Iterable[Mapping]:
-        """
-        :return an iterable containing each record in the response
-        """
-        response_json = response.json()
-        records = response_json.get("data", [])
-        yield from records
-
-
-    @property
-    def primary_key(self) -> Optional[Union[str, List[str], List[List[str]]]]:
-        """
-        :return: string if single primary key, list of strings if composite primary key, list of list of strings if composite primary key consisting of nested fields.
-          If the stream has no primary keys, return None.
-        """
-        return None
-
-
-class Customers(SnowflakeStream):
-    """
-    TODO: Change class name to match the table/data source this stream corresponds to.
-    """
-
-    # TODO: Fill in the primary key. Required. This is usually a unique field in the stream, like an ID or a timestamp.
-    primary_key = "customer_id"
-
-    def path(
-            self, stream_state: Mapping[str, Any] = None, stream_slice: Mapping[str, Any] = None, next_page_token: Mapping[str, Any] = None
-    ) -> str:
-        """
-        TODO: Override this method to define the path this stream corresponds to. E.g. if the url is https://example-api.com/v1/customers then this
-        should return "customers". Required.
-        """
-        return "customers"
-
-
-# Basic incremental stream
-class IncrementalSnowflakeStream(SnowflakeStream, ABC):
-    """
-    TODO fill in details of this class to implement functionality related to incremental syncs for your connector.
-         if you do not need to implement incremental sync for any streams, remove this class.
-    """
-
-    # TODO: Fill in to checkpoint stream reads after N records. This prevents re-reading of data if the stream fails for any reason.
-    state_checkpoint_interval = None
-
-    @property
-    def cursor_field(self) -> str:
-        """
-        TODO
-        Override to return the cursor field used by this stream e.g: an API entity might always use created_at as the cursor field. This is
-        usually id or date based. This field's presence tells the framework this in an incremental stream. Required for incremental.
-
-        :return str: The name of the cursor field.
-        """
-        return []
-
-    def get_updated_state(self, current_stream_state: MutableMapping[str, Any], latest_record: Mapping[str, Any]) -> Mapping[str, Any]:
-        """
-        Override to determine the latest state after reading the latest record. This typically compared the cursor_field from the latest record and
-        the current state and picks the 'most' recent cursor. This is how a stream's state is determined. Required for incremental.
-        """
-        return {}
-
-
-class Employees(IncrementalSnowflakeStream):
-    """
-    TODO: Change class name to match the table/data source this stream corresponds to.
-    """
-
-    # TODO: Fill in the cursor_field. Required.
-    cursor_field = "start_date"
-
-    # TODO: Fill in the primary key. Required. This is usually a unique field in the stream, like an ID or a timestamp.
-    primary_key = "employee_id"
-
-    def path(self, **kwargs) -> str:
-        """
-        TODO: Override this method to define the path this stream corresponds to. E.g. if the url is https://example-api.com/v1/employees then this should
-        return "single". Required.
-        """
-        return "employees"
-
-    def stream_slices(self, stream_state: Mapping[str, Any] = None, **kwargs) -> Iterable[Optional[Mapping[str, any]]]:
-        """
-        TODO: Optionally override this method to define this stream's slices. If slicing is not needed, delete this method.
-
-        Slices control when state is saved. Specifically, state is saved after a slice has been fully read.
-        This is useful if the API offers reads by groups or filters, and can be paired with the state object to make reads efficient. See the "concepts"
-        section of the docs for more information.
-
-        The function is called before reading any records in a stream. It returns an Iterable of dicts, each containing the
-        necessary data to craft a request for a slice. The stream state is usually referenced to determine what slices need to be created.
-        This means that data in a slice is usually closely related to a stream's cursor_field and stream_state.
-
-        An HTTP request is made for each returned slice. The same slice can be accessed in the path, request_params and request_header functions to help
-        craft that specific request.
-
-        For example, if https://example-api.com/v1/employees offers a date query params that returns data for that particular day, one way to implement
-        this would be to consult the stream state object for the last synced date, then return a slice containing each date from the last synced date
-        till now. The request_params function would then grab the date from the stream_slice and make it part of the request by injecting it into
-        the date query param.
-        """
-        raise NotImplementedError("Implement stream slices or delete this method!")
-
 
 # Source
 class SourceSnowflake(AbstractSource):
+
+    SNOWFLAKE_URL_SUFFIX = ".snowflakecomputing.com"
+    HTTP_PREFIX = "https://"
+
     def check_connection(self, logger, config) -> Tuple[bool, any]:
         """
-        TODO: Implement a connection check to validate that the user-provided config can be used to connect to the underlying API
-
-        See https://github.com/airbytehq/airbyte/blob/master/airbyte-integrations/connectors/source-stripe/source_stripe/source.py#L232
-        for an example.
+        TODO: Implement connection using oAyth2.0
+        Done: Connection use JWT token
 
         :param config:  the user-input config object conforming to the connector's spec.yaml
         :param logger:  logger object
         :return Tuple[bool, any]: (True, None) if the input config can be used to connect to the API successfully, (False, error) otherwise.
         """
 
-        snow_flake_url_suffix = ".snowflakecomputing.com"
-        http_prefix = "https://"
         host = config['host']
 
-        if not host.endswith(snow_flake_url_suffix):
-            error_message = f"your host is not ending with {snow_flake_url_suffix}"
-            return (False, error_message)
+        if not host.endswith(self.SNOWFLAKE_URL_SUFFIX):
+            error_message = f"your host is not ending with {self.SNOWFLAKE_URL_SUFFIX}"
+            return False, error_message
 
         url_base = host
 
-        if not url_base.startswith(http_prefix):
-            url_base = f"{http_prefix}{url_base}"
+        if not url_base.startswith(self.HTTP_PREFIX):
+            url_base = f"{self.HTTP_PREFIX}{url_base}"
 
         authenticator = SnowflakeJwtAuthenticator.from_config(config)
 
         stream = CheckConnectionStream(url_base=url_base, config=config, authenticator=authenticator)
         try:
             records = stream.read_records(sync_mode=SyncMode.full_refresh)
-        except Exception as e:
-            return False, e.__str__()
-        for r in records:
-            print('record', r)
+            next(records)
+        except StopIteration:
+            error_message = "There is no stream available for the connection specification provided"
+            return False, error_message
+        except requests.exceptions.HTTPError as error:
+            error_message = error.__str__()
+            error_code = error.args[0]
+            if error_code == 412:
+                error_message = ("SQL execution error for check.\n"
+                                 "The origin of the error is very likely to be:\n"
+                                 "- The configuration provided does not have enough permissions to access the requested database/schema.\n"
+                                 "- The configuration is not consistent (example: schema not present is database.")
+            return False, error_message
+        except Exception as error:
+            return False, error.__str__()
+
         return True, None
 
     def streams(self, config: Mapping[str, Any]) -> List[Stream]:
         """
-        TODO: Replace the streams below with your own streams.
-
         :param config: A Mapping of the user input configuration as defined in the connector spec.
         """
-        # TODO remove the authenticator if not required.
-        auth = TokenAuthenticator(token="api_key")  # Oauth2Authenticator is also available if you need oauth support
-        return [Customers(authenticator=auth), Employees(authenticator=auth)]
+        return []
 
 
-class CheckConnectionStream(SnowflakeStream):
 
-    def __init__(self, url_base, config, **kwargs):
-        super().__init__(**kwargs)
-        self._url_base = url_base
-        self.config = config
-
-    def path(
-            self, stream_state: Mapping[str, Any] = None, stream_slice: Mapping[str, Any] = None, next_page_token: Mapping[str, Any] = None
-    ) -> str:
-        """
-            path of request
-        """
-
-        return f"{self.url_base}/{self.url_suffix}"
-
-    @property
-    def url_base(self):
-        return self._url_base
-
-    @property
-    def statement(self):
-        return "SELECT 1;"
-
-    def request_body_json(
-            self,
-            stream_state: Optional[Mapping[str, Any]],
-            stream_slice: Optional[Mapping[str, Any]] = None,
-            next_page_token: Optional[Mapping[str, Any]] = None,
-    ) -> Optional[Mapping[str, Any]]:
-        json_payload = {
-            "statement": self.statement,
-            "role": self.config['role'],
-            "warehouse": self.config['warehouse'],
-            "database": self.config['database'],
-            "timeout": "1000",
-        }
-        if self.config['schema']:
-            json_payload['schema'] = self.config['schema']
-        print("json_payload", json_payload)
-        return json_payload
